Use sched_yield to wait for suspend barriers
This assumes that entering the kernel scheduler more frequently
improves our chances of noticing suspend requests.

This also increases the number of required timeouts from 5 to
10 in a row, so that we can switch between sched_yield and futex
more frequently.

Move UnignedDifferencei() to a more public place, so that we can
use it more consistently. Require both arguments to have the
same type to avoid any potential comparison conversion issues.

Bug: 97973401
Test: TreeHugger

+++ runtime/gc/heap.cc
@@ -3792,12 +3792,13 @@ void Heap::GrowForUtilization(collector::GarbageCollector* collector_ran,
           current_gc_iteration_.GetFreedRevokeBytes();
       // Records the number of bytes allocated at the time of GC finish,excluding the number of
       // bytes allocated during GC.
-      num_bytes_alive_after_gc_ = UnsignedDifference(bytes_allocated_before_gc, freed_bytes);
+      num_bytes_alive_after_gc_ =
+          UnsignedDifference(static_cast<uint64_t>(bytes_allocated_before_gc), freed_bytes);
       // Bytes allocated will shrink by freed_bytes after the GC runs, so if we want to figure out
       // how many bytes were allocated during the GC we need to add freed_bytes back on.
       // Almost always bytes_allocated + freed_bytes >= bytes_allocated_before_gc.
-      const size_t bytes_allocated_during_gc =
-          UnsignedDifference(bytes_allocated + freed_bytes, bytes_allocated_before_gc);
+      const size_t bytes_allocated_during_gc = UnsignedDifference(
+          bytes_allocated + freed_bytes, static_cast<uint64_t>(bytes_allocated_before_gc));
       // Calculate when to perform the next ConcurrentGC.
       // Estimate how many remaining bytes we will have when we need to start the next GC.
       size_t remaining_bytes = bytes_allocated_during_gc;
